+++ tests/combPRtest.py
@@ -4,12 +4,9 @@ import os, sys
 from PTds import PTdataStore
 from Resource import Resource
 from ResourceIndex import ResourceIndex
+from PassFail import PassFail
 
-
-
-
-
-def test1(ptds):
+def test1(ptds, pf):
     # test retrieval of simple performance results by context. 
     # retrieve ancestors and descendants
     # Assumes that there are executions with names: 
@@ -22,12 +19,12 @@ def test1(ptds):
         reses = ptds.getPerfResultsByContext([(ex,t)], anc=True, desc=True, \
                      combined=False)
         if len(reses) != pr_count:
-           print "FAIL: for %s. Got %s, expected %s" %(ex,len(reses),pr_count)
+            pf.failed("for %s. Got %s, expected %s" %(ex,len(reses),pr_count))
         else:
-           print "PASS"
+            pf.passed()
         
         
-def test2(ptds):
+def test2(ptds, pf):
     # test retrieval of simple performance results by context. 
     # retrieve ancestors, but not descendants
     # Assumes that there are executions with names: 
@@ -41,13 +38,13 @@ def test2(ptds):
         results = ptds.getPerfResultsByContext([(r,t)], anc=True, desc=False, \
                      combined=False)
         if len(results) != pr_count:
-           print "FAIL: for %s. Got %s, expected %s" \
-                 % (r,len(results),pr_count)
+            pf.failed("For %s. Got %s, expected %s" % (r,len(results),pr_count))
+                
         else:
-           print "PASS"
+            pf.passed()
         
     
-def test3(ptds):
+def test3(ptds, pf):
     # test retrieval of simple performance results by context. 
     # retrieve descendants, but not ancestors 
     # Assumes that there are executions with names: 
@@ -60,12 +57,11 @@ def test3(ptds):
         results = ptds.getPerfResultsByContext([(r,t)], anc=False, desc=True, \
                      combined=False)
         if len(results) != pr_count:
-           print "FAIL: for %s. Got %s, expected %s" \
-                 % (r,len(results),pr_count)
+            pf.failed("for %s. Got %s, expected %s" % (r,len(results),pr_count))
         else:
-           print "PASS"
+            pf.passed()
 
-def test4(ptds):
+def test4(ptds, pf):
     # test retrieval of performance results by context specifying combined, 
     # when they are simple perf reses
     # retrieve ancestors and descendants
@@ -79,11 +75,11 @@ def test4(ptds):
         reses = ptds.getPerfResultsByContext([(ex,t)], anc=True, desc=True, \
                      combined=True)
         if len(reses) != pr_count:
-           print "FAIL: for %s. Got %s, expected %s" %(ex,len(reses),pr_count)
+            pf.failed("for %s. Got %s, expected %s" %(ex,len(reses),pr_count))
         else:
-           print "PASS"
+            pf.passed()
         
-def test5(ptds):
+def test5(ptds, pf):
     # test retrieval of performance results by label 
     # when they are simple perf reses
     # retrieve ancestors and descendants
@@ -104,11 +100,11 @@ def test5(ptds):
     ex = '/sppm-2'
     reses = ptds.getPerfResultsByLabel(testLabel, combined=False)
     if len(reses) != pr_count:
-       print "FAIL: for %s. Got %s, expected %s" %(ex,len(reses),pr_count)
+        pf.failed("For %s Got %s, expected %s" %(ex,len(reses),pr_count))
     else:
-       print "PASS"
+        pf.passed()
  
-def test6(ptds):
+def test6(ptds, pf):
     # test create combined context
     # retrieves several contexts from the db and creates a combined context
     # from them
@@ -135,19 +131,20 @@ def test6(ptds):
         fids.append(fid)
     newCntxt = ptds.createCombinedContext(fids)    
     if newCntxt == expectedContext:
-       print "PASS"
+        pf.passed()
     else:
-       print "FAIL: unexpected context for TSTEP-HYD cpu time"
+        pf.failed("unexpected context for TSTEP-HYD cpu")
+
     fids = [] 
     for id,met,val,units,st,et,fid,lab,comb in ts_wall:
         fids.append(fid)
     newCntxt = ptds.createCombinedContext(fids)    
     if newCntxt == expectedContext:
-       print "PASS"
+        pf.passed()
     else:
-       print "FAIL: unexpected context for TSTEP-HYD wall time"
+        pf.failed("unexpected context for TSTEP-HYD wall time")
 
-def test7(ptds):
+def test7(ptds, pf):
     # test create combined perf result 
     # retrieves several simple prs from the db and creates a combined pr  
     # from them
@@ -179,29 +176,33 @@ def test7(ptds):
         fids.append(fid)
         prIds.append(id)
         totVal += val
+    if len(ts_cpu) == 0:
+        pf.failed("TS CPU has no length")
+        return
+
     newVal = totVal/len(ts_cpu)
     newCntxt = ptds.createCombinedContext(fids)
     ret = ptds.addResource("/Average TSTEP-HYD cpu time", "metric")
     if ret == 0:
-       print "FAIL: adding metric 'Average TSTEP-HYD cpu time' failed"
+        pf.failed("adding metric 'Average TSTEP-HYD cpu time' failed")
     cprid = ptds.addCombinedPerfResult(newCntxt, "/Average TSTEP-HYD cpu time", newVal, "seconds", "noValue", "noValue", prIds, label=newLabel)
     if cprid == 0:
-       print "FAIL: addCombinedPerfResult failed for Average TSTEP-HYD cpu time"
+        pf.failed("addCombinedPerfResult failed for Average TSTEP-HYD cpu time")
     cpr = ptds.getPerfResultsByLabel(newLabel, combined=True)
     if len(cpr) != 1:
-       print "FAIL: wrong number of results in db for label=%s for Average TSTEP-HYD cpu time" % newLabel
+        pf.failed("wrong number of results in db for label=%s for Average TSTEP-HYD cpu time" % newLabel)
     expectedContext = "/sppm,/build,/mpiicc-8.0 Version 8.0,/mpiifort-8.0 Version 8.0,/env,/sppm-2/Process-0,/SingleMachineMCR/MCR,/inputdeck,/iq.h,/Linux #1 SMP Mon Sep 27 13:51:13 PDT 2004 2.4.21-p4smp-71.3chaos,/Linux #1 SMP Wed Sep 1 16:37:16 PDT 2004 2.4.21-p4smp-71chaos,/fpp,/m4,/whole execution/10,/whole execution/12,/whole execution/14,/whole execution/16,/whole execution/18,/whole execution/2,/whole execution/20,/whole execution/4,/whole execution/6,/whole execution/8"
     # extract the focus id from the combined perf res we made above
     id,met,val,units,st,et,fid,lab,comb = cpr[0]
     # check to see that there's a new focus for our new perf result
     fidFromDb = ptds.findFocusByName(expectedContext)
     if fidFromDb != fid:
-       print "FAIL: wrong context entered for Average TSTEP-HYD cpu time"
+        pf.failed("wrong context entered for Average TSTEP-HYD cpu time")
     ret = ptds.getCombinedPerfResultSourceData(cprid)
     if len(ret) != len(ts_cpu):
-       print "FAIL: wrong number of source data prs in db for Average TSTEP-HYD cpu time" 
+        pf.failed("wrong number of source data prs in db for Average TSTEP-HYD cpu time")
     else:
-       print "PASS"
+        pf.passed()
 
 
     # average the values of TSTEP-HYD wall time
@@ -216,18 +217,18 @@ def test7(ptds):
     newCntxt = ptds.createCombinedContext(fids)
     ret = ptds.addResource("/Average TSTEP-HYD wall time", "metric")
     if ret == 0:
-       print "FAIL: adding metric 'Average TSTEP-HYD wall time' failed"
+       pf.failed("adding metric 'Average TSTEP-HYD wall time' failed")
     cprid = ptds.addCombinedPerfResult(newCntxt, "/Average TSTEP-HYD wall time", newVal, "seconds", "noValue", "noValue", prIds, label=newLabel)
     if cprid == 0:
-      print "FAIL: addCombinedPerfResult failed for Average TSTEP-HYD wall time"
+        pf.failed("addCombinedPerfResult failed for Average TSTEP-HYD wall time")
     ret = ptds.getPerfResultsByLabel(newLabel, combined=True)
     if len(ret) != 2:
-       print "FAIL: wrong number of results in db for label=%s for Average TSTEP-HYD wall time" % newLabel
+       pf.failed("wrong number of results in db for label=%s for Average TSTEP-HYD wall time" % newLabel)
     ret = ptds.getCombinedPerfResultSourceData(cprid)
     if len(ret) != len(ts_wall):
-       print "FAIL: wrong number of source data prs in db for Average TSTEP-HYD wall time" 
+        pf.failed("wrong number of source data prs in db for Average TSTEP-HYD wall time")
     else:
-       print "PASS"
+        pf.passed()
 
     # now to test creating a combined res from combined reses
     # please note: this actual combination of averaging cpu and wall time
@@ -245,21 +246,21 @@ def test7(ptds):
     newCntxt = ptds.createCombinedContext(fids)
     ret = ptds.addResource("/Average of TSTEP-HYD wall and cpu time", "metric")
     if ret == 0:
-       print "FAIL: adding metric 'Average of TSTEP-HYD wall and cpu time' failed"
+       pf.failed("adding metric 'Average of TSTEP-HYD wall and cpu time' failed")
     cprid = ptds.addCombinedPerfResult(newCntxt, "/Average of TSTEP-HYD wall and cpu time", newVal, "seconds", "noValue", "noValue", prIds, label=newLabel)
     if cprid == 0:
-      print "FAIL: addCombinedPerfResult failed for Average of TSTEP-HYD wall and cpu time"
+        pf.failed("addCombinedPerfResult failed for Average of TSTEP-HYD wall and cpu time")
     ret = ptds.getPerfResultsByLabel(newLabel, combined=True)
     if len(ret) != 3:
-       print "FAIL: wrong number of results in db for label=%s for Average of TSTEP-HYD wall  and cpu time" % newLabel
+        pf.failed("wrong number of results in db for label=%s for Average of TSTEP-HYD wall  and cpu time" % newLabel)
     ret = ptds.getCombinedPerfResultSourceData(cprid)
     if len(ret) != len(prs):
-       print "FAIL: wrong number of source data prs in db for Average of TSTEP-HYD wall and cpu time" 
+        pf.failed("wrong number of source data prs in db for Average of TSTEP-HYD wall and cpu time")
     else:
-       print "PASS"
+        pf.passed()
    
 
-def test8(ptds):
+def test8(ptds,pf):
     # test delete combined perf result 
     # retrieves combined prs from the db and tries to delete them
     # Assumes that test7 has been run already
@@ -270,14 +271,14 @@ def test8(ptds):
     newLabel = 'sppm-2'
     ret = ptds.getPerfResultsByLabel(newLabel, combined=True)
     if not ret:
-       print "FAIL: perhaps test7 wasn't run yet? the expected data is not in the database"
+       pf.failed("perhaps test7 wasn't run yet? the expected data is not in the database")
     numStarted = len(ret)
     numDeleted = 0
     idsNotDeleted = []
     for id,met,val,units,st,et,fid,lab,comb in ret:
         ret,list = ptds.deleteCombinedPerfResult(id)
         if ret == 1 and len(list) > 0:
-           print "FAIL: deleted comb perf rest with dependencies"
+            pf.failed("deleted comb perf rest with dependencies")
         elif ret == 1 and len(list) == 0:
            numDeleted += 1
         elif ret == 0 and len(list) > 0:
@@ -287,30 +288,33 @@ def test8(ptds):
            print "FAIL: unexpected return value from deleteCombinedPerfResult"
     ret = ptds.getPerfResultsByLabel(newLabel, combined=True)
     if len(ret) != (numStarted-numDeleted):
-       print "FAIL: retrieve unexpected number of prs after deletion." 
+        pf.failed("retrieve unexpected number of prs after deletion.")
   
     # the list idsNotDeleted contains 2 combined perf reses that share a 
     # context (from test7). 
     #we'll delete one, and make sure the context is still there, and then
     # delete the other and make sure the context is gone
     expectedContext = "/sppm,/build,/mpiicc-8.0 Version 8.0,/mpiifort-8.0 Version 8.0,/env,/sppm-2/Process-0,/SingleMachineMCR/MCR,/inputdeck,/iq.h,/Linux #1 SMP Mon Sep 27 13:51:13 PDT 2004 2.4.21-p4smp-71.3chaos,/Linux #1 SMP Wed Sep 1 16:37:16 PDT 2004 2.4.21-p4smp-71chaos,/fpp,/m4,/whole execution/10,/whole execution/12,/whole execution/14,/whole execution/16,/whole execution/18,/whole execution/2,/whole execution/20,/whole execution/4,/whole execution/6,/whole execution/8"
+    if (len(idsNotDeleted) == 0):
+        pf.failed("Missing idsNotDeleted")
+        return
+
     id,fid = idsNotDeleted[0]
     ret,list = ptds.deleteCombinedPerfResult(id) 
     if ret!=0 and len(list) != 0:
-       print "FAIL: did not delete pr with no dependencies, #1"
+        pf.failed("did not delete pr with no dependencies, #1")
     fidFromDB = ptds.findFocusByName(expectedContext)
     if fidFromDB != fid:
-       print "FAIL: expected to find context but it was gone"
+        pf.failed("expected to find context but it was gone")
     id,fid = idsNotDeleted[1]
     ret,list = ptds.deleteCombinedPerfResult(id) 
     if ret!=0 and len(list) != 0:
-       print "FAIL: did not delete pr with no dependencies, #2"
+        pf.failed("did not delete pr with no dependencies, #2")
     fidFromDB = ptds.findFocusByName(expectedContext)
     if fidFromDB == fid:
-       print "FAIL: found context when it should be gone"
-    
+        pf.failed("found context when it should be gone")
     else:
-       print "PASS"
+        pf.passed()
 
     
 def main():
@@ -323,9 +327,12 @@ def main():
 
     tests = [test1,test2,test3,test4,test5,test6,test7,test8]
 
+    pf = PassFail()
     for test in tests: 
-        test(ptds)
+        test(ptds, pf)
 
+    pf.test_info()
+    return pf.failed_count > 0
 
 if __name__ == "__main__":
    sys.exit(main())
